コードレビュー修正2/27 分でgoodスタンプを押した部分
