--- Colombia/Sources/Extension/UICollectionView+Ex.swift
@@ -1,15 +1,0 @@
-//
-//  UICollectionView+Ex.swift
-//  Colombia
-//
-//  Created by 化田晃平 on R 3/02/26.
-//
-
-import UIKit
-extension UICollectionView {
-    var showingRowNum: Int {
-        get {
-            return 3
-        }
-    }
-}
