+++ Colombia/Sources/Repository/AnnictDataRepository.swift
@@ -12,11 +12,15 @@ protocol Repository {
     associatedtype Response
     var apiClient: APIClient { get }
     func fetch() -> Observable<Response>
+    
 }
 
 struct AnnictDataRepository: Repository {
+    
     let apiClient = APIClient()
+    
     typealias Response = AnnictAPIModel
+    
     func fetch() -> Observable<AnnictAPIModel> {
         let request = AnnictAPIRequest(endpoint: .works)
         return apiClient.request(request)
